bugfix: check if thread is suspended or not in :Wait
If you try firing a signal that's waiting but on a thread that was `coroutine.close`d then an error happens (only when fired with args...)

In a way it's good because it's a nice warning. Errors aren't negative necessarily. Sometimes they're just expected. I'd like to either completely ignore or be even more strict and always error if the coroutine is closed.